pyproject.toml
[project]
name = "nfs-ansible-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ansible>=12.0.0",
    "ansible-creator>=25.8.0",
]

# pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "integration",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-testinfra>=10.2.2",
]
